Propagate connection lost errors to crate clients (#1)
Allow for crate users to get error when fluentd connection is lost so they can handle this case. This is a first small step towards this crate handling connection lost: ideally I'd like the crate to restore the connection by itself (based on provided configuration), but that would take more work.